Moved cwMax to earlier in the file for clarity
The bulb has a max power capacity. Through experimentation, it appears to be 3 lights at full power, though it's difficult to be sure. we use cold and warm combined at half power as our full-white value, but the Wiz app uses only warm at 140/255 as its white. This doesn't look right when using homekit to set color temperature, because it assumes only RGB bulbs are involved in the color production.